requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
Django==4.0.1
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
django-mysql==4.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
Markdown==3.3.6
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
python-decouple==3.5
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
stripe==2.65.0
urllib3==1.26.8
zipp==3.7.0